完成二课界面

+++ app/src/main/java/com/thryan/secondclass/core/result/ScoreInfo.kt
@@ -9,6 +9,6 @@ import java.math.BigDecimal
  * @param activity 完成活动数
  */
 data class ScoreInfo(
-    val score: BigDecimal, val item: Int, val integrity_value: Int,
+    val score: Double, val item: Int, val integrity_value: Int,
     val activity: Int
 )
